correction heredoc et pipe
correction des heredoc qui ne doit pas s'ouvrir (pipe restant ouvert)
correction des pipefd qui ne se lancer quand cas de infile juste (fermeture de pipe impossible -> segfault)

+++ src/execution/heredoc_handler.c
@@ -6,7 +6,7 @@
 /*   By: njegat <njegat@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/04 21:58:24 by njegat            #+#    #+#             */
-/*   Updated: 2023/04/27 06:51:13 by njegat           ###   ########.fr       */
+/*   Updated: 2023/05/11 22:25:50 by njegat           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -76,6 +76,11 @@ int	heredoc_handler(char *limiter)
 	{
 		heredoc_fd(fd, limiter);
 		close(fd);
+	}
+	if (exitcode()->write_by_signale == TRUE)
+	{
+		free(rand_name);
+		return (0);
 	}
 	fd = open(rand_name, O_RDONLY, 0644);
 	unlink(rand_name);
